requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pypdf2
chromadb
langchain-text-splitters
python-docx
python-pptx
markdown